rename test to spec
