+++ modules/work-dictionary/src/test/java/com/kgmyshin/workDictionary/ui/work/popular/PopularWorkListPresenterSpec.kt
@@ -1,4 +1,4 @@
-package com.kgmyshin.workDictionary.ui.work.season
+package com.kgmyshin.workDictionary.ui.work.popular
 
 import com.kgmyshin.common.errorHandler.ErrorHandler
 import com.kgmyshin.workDictionary.domain.DomainHelper
@@ -6,7 +6,7 @@ import com.kgmyshin.workDictionary.ui.work.ScreenTransition
 import com.kgmyshin.workDictionary.ui.work.WorkListContract
 import com.kgmyshin.workDictionary.ui.work.WorkViewModelConverter
 import com.kgmyshin.workDictionary.ui.work.WorkViewModelFactory
-import com.kgmyshin.workDictionary.usecase.GetBeforeSeasonWorkListUseCase
+import com.kgmyshin.workDictionary.usecase.GetPopularWorkListUseCase
 import io.reactivex.Single
 import io.reactivex.schedulers.Schedulers
 import org.jetbrains.spek.api.dsl.given
@@ -18,16 +18,16 @@ import org.junit.runner.RunWith
 import org.mockito.Mockito
 
 @RunWith(JUnitPlatform::class)
-internal class BeforeSeasonWorkListPresenterTest : SubjectSpek<BeforeSeasonWorkListPresenter>({
+internal class PopularWorkListPresenterSpec : SubjectSpek<PopularWorkListPresenter>({
 
-    val getBeforeSeasonWorkListUseCase = Mockito.mock(GetBeforeSeasonWorkListUseCase::class.java)
+    val getPopularWorkListUseCase = Mockito.mock(GetPopularWorkListUseCase::class.java)
     val errorHandler = Mockito.mock(ErrorHandler::class.java)
     val view = Mockito.mock(WorkListContract.View::class.java)
     val screenTransition = Mockito.mock(ScreenTransition::class.java)
 
     subject {
-        BeforeSeasonWorkListPresenter(
-                getBeforeSeasonWorkListUseCase,
+        PopularWorkListPresenter(
+                getPopularWorkListUseCase,
                 Schedulers.trampoline(),
                 errorHandler
         ).apply {
@@ -38,17 +38,16 @@ internal class BeforeSeasonWorkListPresenterTest : SubjectSpek<BeforeSeasonWorkL
         }
     }
 
-    given("GetBeforeSeasonWorkListUseCase return workList") {
+    given("GetPopularWorkListUseCase return workList") {
 
         val workList = listOf(
                 DomainHelper.work(),
                 DomainHelper.work(),
                 DomainHelper.work()
         )
-        Mockito.`when`(getBeforeSeasonWorkListUseCase.execute()).thenReturn(Single.just(workList))
+        Mockito.`when`(getPopularWorkListUseCase.execute()).thenReturn(Single.just(workList))
 
         on("onCreateView") {
-
             subject.onCreateView()
 
             it("should setUp ViewModel to view") {
@@ -58,8 +57,8 @@ internal class BeforeSeasonWorkListPresenterTest : SubjectSpek<BeforeSeasonWorkL
                 Mockito.verify(view).dismissProgress()
             }
         }
-
     }
+
 
     given("") {
