+++ modules/work-dictionary/src/test/java/com/kgmyshin/workDictionary/ui/work/search/SearchWorkListPresenterSpec.kt
@@ -1,12 +1,12 @@
-package com.kgmyshin.workDictionary.ui.work.season
+package com.kgmyshin.workDictionary.ui.work.search
 
 import com.kgmyshin.common.errorHandler.ErrorHandler
+import com.kgmyshin.random.RandomHelper
 import com.kgmyshin.workDictionary.domain.DomainHelper
 import com.kgmyshin.workDictionary.ui.work.ScreenTransition
-import com.kgmyshin.workDictionary.ui.work.WorkListContract
 import com.kgmyshin.workDictionary.ui.work.WorkViewModelConverter
 import com.kgmyshin.workDictionary.ui.work.WorkViewModelFactory
-import com.kgmyshin.workDictionary.usecase.GetThisSeasonWorkListUseCase
+import com.kgmyshin.workDictionary.usecase.SearchWorkListUseCase
 import io.reactivex.Single
 import io.reactivex.schedulers.Schedulers
 import org.jetbrains.spek.api.dsl.given
@@ -18,16 +18,16 @@ import org.junit.runner.RunWith
 import org.mockito.Mockito
 
 @RunWith(JUnitPlatform::class)
-internal class ThisSeasonWorkListPresenterTest : SubjectSpek<ThisSeasonWorkListPresenter>({
+internal class SearchWorkListPresenterSpec : SubjectSpek<SearchWorkListPresenter>({
 
-    val getThisSeasonWorkListUseCase = Mockito.mock(GetThisSeasonWorkListUseCase::class.java)
+    val searchWorkListUseCase = Mockito.mock(SearchWorkListUseCase::class.java)
     val errorHandler = Mockito.mock(ErrorHandler::class.java)
-    val view = Mockito.mock(WorkListContract.View::class.java)
+    val view = Mockito.mock(SearchWorkListContract.View::class.java)
     val screenTransition = Mockito.mock(ScreenTransition::class.java)
 
     subject {
-        ThisSeasonWorkListPresenter(
-                getThisSeasonWorkListUseCase,
+        SearchWorkListPresenter(
+                searchWorkListUseCase,
                 Schedulers.trampoline(),
                 errorHandler
         ).apply {
@@ -38,18 +38,17 @@ internal class ThisSeasonWorkListPresenterTest : SubjectSpek<ThisSeasonWorkListP
         }
     }
 
-    given("GetThisSeasonWorkListUseCase return workList") {
-
+    given("SearchWorkListUseCase return workList") {
+        val keyword = RandomHelper.randomString()
         val workList = listOf(
                 DomainHelper.work(),
                 DomainHelper.work(),
                 DomainHelper.work()
         )
-        Mockito.`when`(getThisSeasonWorkListUseCase.execute()).thenReturn(Single.just(workList))
+        Mockito.`when`(searchWorkListUseCase.execute(keyword)).thenReturn(Single.just(workList))
 
-        on("onCreateView") {
-
-            subject.onCreateView()
+        on("onUpdateKeyword") {
+            subject.onUpdateKeyword(keyword)
 
             it("should setUp ViewModel to view") {
                 val expected = WorkViewModelConverter.convertToViewModel(workList)
@@ -58,7 +57,6 @@ internal class ThisSeasonWorkListPresenterTest : SubjectSpek<ThisSeasonWorkListP
                 Mockito.verify(view).dismissProgress()
             }
         }
-
     }
 
     given("") {
